do plan

+++ broker-api/broker-console-api/src/main/java/org/limbo/flowjob/broker/api/console/param/DispatchOptionParam.java
@@ -23,6 +23,8 @@ import lombok.Data;
 import org.limbo.flowjob.broker.api.constants.enums.DispatchType;
 import org.limbo.flowjob.broker.api.constants.enums.LoadBalanceType;
 
+import java.math.BigDecimal;
+
 /**
  * @author Devil
  * @since 2021/7/24
@@ -53,12 +55,12 @@ public class DispatchOptionParam {
      * 所需的CPU核心数，小于等于0表示此作业未定义CPU需求。在分发作业时，会根据此方法返回的CPU核心需求数量来检测一个worker是否有能力执行此作业。
      */
     @Schema(title = "所需的CPU核心数", description = "小于等于0表示此作业未定义CPU需求")
-    private Float cpuRequirement;
+    private BigDecimal cpuRequirement;
 
     /**
      * 所需的内存GB数，小于等于0表示此作业未定义内存需求。在分发作业时，会根据此方法返回的内存需求数量来检测一个worker是否有能力执行此作业。
      */
     @Schema(title = "所需的内存GB数", description = "小于等于0表示此作业未定义内存需求")
-    private Float ramRequirement;
+    private BigDecimal ramRequirement;
 
 }
